+++ src/EtcdTerminal.App/CHANGELOG.md
@@ -12,6 +12,10 @@
 - Removed "Select etcd instance:" header; now shows helpful message to go to Manage Connections when no connections exist
 - Cursor is hidden where not required
 
+### Fixed
+
+- Reset keys list after manipulations
+
 ## [0.5] - 2026-08-13
 
 ### Added